move states to states.js file
