+++ src/states.js
@@ -1,0 +1,8 @@
+var states = {
+  "HANDSHAKING": "handshaking",
+  "STATUS": "status",
+  "LOGIN": "login",
+  "PLAY": "play"
+};
+
+module.exports=states;
